+++ src/main/java/students/student_svetlana_seda/lesson_3/level_1/Robot.java
@@ -1,19 +1,25 @@
-package students.student_svetlana_seda.level_1;
+package students.student_svetlana_seda.lesson_3.level_1;
 
 public class Robot {
     String name;
+
     public void sayHello() {
         System.out.println("Hello!");
     }
-        Robot(String name) {
+
+    Robot(String name) {
         this.name = name;
-        }
+    }
+
     public void sayYourName() {
         System.out.println("My name is " + this.name + ".");
     }
-    public void sayByeBye(){
+
+    public void sayByeBye() {
         System.out.println("Bye!");
     }
-    public Robot(){
+
+    public Robot() {
     }
+
 }
