--- src/main/java/students/student_svetlana_seda/level_1/RobotDemo.java
@@ -1,19 +1,0 @@
-package students.student_svetlana_seda.level_1;
-public class RobotDemo {
-    public static void main(String[] args) {
-     Robot firstRobot = new Robot ("Gold");
-     Robot secondRobot = new Robot ("Silver");
-     Robot thirdRobot = new Robot ("Metal");
-firstRobot.sayHello();
-firstRobot.sayYourName();
-firstRobot.sayByeBye();
-        System.out.println("___________");
-secondRobot.sayHello();
-secondRobot.sayYourName();
-secondRobot.sayByeBye();
-        System.out.println("---------------");
-thirdRobot.sayHello();
-thirdRobot.sayYourName();
-thirdRobot.sayByeBye();
-    }
-}
